VyasaAi-backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.13
python-dotenv==1.0.1
oracledb==1.3.2
requests==2.31.0
